feat(lsp): add rust_analyzer and update tinymist setup
This commit introduces the setup for rust_analyzer in the lsp
configuration. It enables inlay hints and sets the checkOnSave command
to "clippy". It also enables procMacro.

In addition, the setup for tinymist is updated to support single file
and the root directory is set to the current working directory.